+++ skyvern/forge/sdk/routes/__init__.py
@@ -1,4 +1,4 @@
 from skyvern.forge.sdk.routes import agent_protocol  # noqa: F401
 from skyvern.forge.sdk.routes import browser_sessions  # noqa: F401
+from skyvern.forge.sdk.routes import credentials  # noqa: F401
 from skyvern.forge.sdk.routes import streaming  # noqa: F401
-from skyvern.forge.sdk.routes import totp  # noqa: F401
